adjusted code removed unwanted code updated bookmark
