+++ newsaggregator/templates/components/navbar.html
@@ -25,7 +25,7 @@
                 <a href="{% url 'core:index' %}" class="nav-link">Home</a>
             </li>
             <li class="nav-item">
-                <a href="{% url 'news:view_bookmarks' %}" class="nav-link">My Bookmarks</a>
+                <a href="{% url 'core:view_bookmarks' %}" class="nav-link">My Bookmarks</a>
             </li>
             <li>
                 <a href="{% url 'core:about' %}" class="nav-link">About Us</a>
@@ -63,7 +63,7 @@
                     <a href="{% url 'core:index' %}" class="nav-link">Home</a>
                 </li>
                 <li class="nav-item">
-                    <a href="{% url 'news:view_bookmarks' %}" class="nav-link">My Bookmarks</a>
+                    <a href="{% url 'core:view_bookmarks' %}" class="nav-link">My Bookmarks</a>
                 </li>
 
                 <li class="nav-item">
