+++ newsaggregator/templates/components/swiperjs.html
@@ -38,7 +38,7 @@
 <!-- Swiper -->
 <div class="swiper mySwiper">
     <div class="swiper-wrapper">
-        {% for record in swiper %}
+        {% for record in object_list %}
         <div class="col-12 col-md-6 col-lg-4 mb-4 newscard swiper-slide">
             <div class="row">
                 <div class="col-md-12 swiper-image">
